requirements.txt
Django==1.7.1
Markdown==2.5.2
argparse==1.2.1
django-filter==0.9.1
djangorestframework==2.4
djangorestframework-gis==0.7
wsgiref==0.1.2
